requirements.txt
asgiref==3.5.2
Django==3.2.14
django-jazzmin==2.5.0
django-jsonfield==1.4.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
